Change `UserProfile::statusExpiration` type to `long`

+++ slack-base/src/main/java/com/hubspot/slack/client/models/users/UserProfileIF.java
@@ -3,9 +3,9 @@ package com.hubspot.slack.client.models.users;
 import java.util.Map;
 import java.util.Optional;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
 import org.immutables.value.Value.Immutable;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.PropertyNamingStrategy.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
@@ -29,7 +29,7 @@ public interface UserProfileIF {
   Optional<String> getEmail();
   Optional<String> getStatusText();
   Optional<String> getStatusEmoji();
-  Optional<Integer> getStatusExpiration();
+  Optional<Long> getStatusExpiration();
   Optional<String> getTitle();
   Optional<String> getPhone();
   Optional<String> getSkype();
